Worked on TechnicalDocumentation and changed imgs on tributepage
Worked on filling the content of TechnicalDocumentation.
added style and javascript.
style should follow the layout of the demo, javascript for smoothing and responsiveness.

changes in tributepage are some images, images used now are taken personally.
portraits are taken from books and news articles.
removed old photos taken from the internet.
changed the lines for the bg images to the correct format as it lacked the correct folder path.